Move WorkflowRepo to separate file
